M2-5585: Extract ActivityGrid as component
