+++ src/modules/Dashboard/components/ActivityGrid/ActivityGrid.styles.ts
