+++ src/modules/Dashboard/features/Applet/Activities/Activities.tsx
@@ -1,37 +1,28 @@
 import { useState, useMemo, useEffect, useCallback } from 'react';
-import { useTranslation } from 'react-i18next';
 import { generatePath, useNavigate, useParams } from 'react-router-dom';
 import { format } from 'date-fns';
 
-import {
-  ActionsMenu,
-  ButtonWithMenu,
-  Spinner,
-  Svg,
-  MenuActionProps,
-  EmptyState,
-  Row,
-} from 'shared/components';
-import { StyledBody, StyledFlexTopCenter, StyledFlexWrap } from 'shared/styles';
+import { ActionsMenu, MenuActionProps, Row } from 'shared/components';
+import { StyledFlexTopCenter } from 'shared/styles';
 import { getAppletActivitiesApi } from 'api';
-import { useAsync, useTable } from 'shared/hooks';
+import { useAsync } from 'shared/hooks';
 import { DateFormats } from 'shared/consts';
 import { page } from 'resources';
 import { Activity, workspaces } from 'redux/modules';
-import { ActivitySummaryCard } from 'modules/Dashboard/components';
-
-import { StyledButton, StyledSearch, StyledSvg } from './Activities.styles';
-import { ActivitiesData, ActivityActionProps } from './Activities.types';
-import { getActivityActions } from './Activities.utils';
+import { ActivityGrid } from 'modules/Dashboard/components/ActivityGrid/ActivityGrid';
+import { StyledSvg } from 'modules/Dashboard/components/ActivityGrid/ActivityGrid.styles';
+import { getActivityActions } from 'modules/Dashboard/components/ActivityGrid/ActivityGrid.utils';
+import {
+  ActivitiesData,
+  ActivityActionProps,
+} from 'modules/Dashboard/components/ActivityGrid/ActivityGrid.types';
 
 export const Activities = () => {
   const { appletId } = useParams();
   const navigate = useNavigate();
-  const { t } = useTranslation('app');
   const [activitiesData, setActivitiesData] = useState<ActivitiesData | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
+  const [isLoading, setIsLoading] = useState(true);
   const dataTestid = 'dashboard-applet-activities';
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
 
   const workspaceRoles = workspaces.useRolesData();
   const roles = appletId ? workspaceRoles?.data?.[appletId] : undefined;
@@ -47,17 +38,19 @@ export const Activities = () => {
     () => setIsLoading(false),
   );
 
-  const { handleSearch, handleReload, searchValue } = useTable(async ({ params }) => {
+  useEffect(() => {
     if (!appletId) return;
 
-    setIsLoading(true);
-    await getActivities({
+    getActivities({
       params: {
-        ...params,
         appletId,
       },
     });
-  });
+
+    return () => {
+      setActivitiesData(null);
+    };
+  }, [appletId, getActivities]);
 
   const actions = useMemo(
     () => ({
@@ -153,123 +146,25 @@ export const Activities = () => {
   );
 
   const activities = useMemo(
-    () => (activitiesData?.result ?? []).map((activity) => formatRow(activity)),
+    () => (activitiesData?.result ?? []).map((activity) => formatRow(activity as Activity)),
     [activitiesData, formatRow],
   );
 
   useEffect(() => {
     if (!appletId) return;
 
-    handleReload();
-
     return () => {
       setActivitiesData(null);
     };
   }, [appletId]);
 
-  const isEmpty = !activities?.length && !isLoading;
-
-  const getEmptyComponent = () => {
-    if (isEmpty) {
-      if (searchValue) {
-        return t('noMatchWasFound', { searchValue });
-      }
-
-      return appletId ? t('noActivitiesForApplet') : t('noActivities');
-    }
-  };
-
   return (
-    <StyledBody>
-      {isLoading && <Spinner />}
-
-      <StyledFlexWrap sx={{ gap: 1.2, mb: 2.4 }}>
-        {appletId && (
-          <>
-            <StyledFlexTopCenter sx={{ gap: 1.2 }}>
-              <StyledButton
-                variant="outlined"
-                startIcon={<Svg width={18} height={18} id="slider-rows" />}
-                // TODO: Implement filters
-                // https://mindlogger.atlassian.net/browse/M2-5530
-                onClick={() => alert('TODO: filters')}
-                data-testid={`${dataTestid}-filters`}
-              >
-                {t('filters')}
-              </StyledButton>
-
-              <ButtonWithMenu
-                variant="outlined"
-                label={t('sortBy')}
-                anchorEl={anchorEl}
-                setAnchorEl={setAnchorEl}
-                // TODO: Implement sorting
-                // https://mindlogger.atlassian.net/browse/M2-5445
-                menuItems={[
-                  {
-                    title: 'TODO: Sort options',
-                    action: () => alert('TODO: Sort options'),
-                  },
-                ]}
-                startIcon={<></>}
-                data-testid={`${dataTestid}-sort-by`}
-              />
-            </StyledFlexTopCenter>
-
-            <StyledFlexWrap sx={{ gap: 1.2, ml: 'auto' }}>
-              <StyledSearch
-                withDebounce
-                // TODO: Implement search
-                // https://mindlogger.atlassian.net/browse/MLG-21
-                placeholder={`TODO: ${t('searchActivities')}`}
-                onSearch={handleSearch}
-                data-testid={`${dataTestid}-search`}
-              />
-
-              <StyledButton
-                // TODO: Replace with missing `tonal` button variant as shown in Figma
-                // https://mindlogger.atlassian.net/browse/M2-6071
-                variant="outlined"
-                // TODO: Implement assign
-                // https://mindlogger.atlassian.net/browse/M2-5710
-                onClick={() => alert('TODO: Assign activity')}
-                data-testid={`${dataTestid}-assign`}
-                sx={{ minWidth: '10rem' }}
-              >
-                {t('assign')}
-              </StyledButton>
-
-              <StyledButton
-                variant="contained"
-                onClick={() => navigate(generatePath(page.builderAppletActivities, { appletId }))}
-                data-testid={`${dataTestid}-add-activity`}
-                sx={{ minWidth: '13.2rem' }}
-              >
-                {t('addActivity')}
-              </StyledButton>
-            </StyledFlexWrap>
-          </>
-        )}
-      </StyledFlexWrap>
-
-      {!!activities.length && (
-        <StyledFlexWrap sx={{ gap: 2.4, overflowY: 'auto' }} data-testid={`${dataTestid}-grid`}>
-          {activities.map((activity, index) => (
-            <ActivitySummaryCard
-              key={index}
-              name={String(activity.name.value)}
-              image={String(activity.image.value)}
-              actionsMenu={activity.actions.content()}
-              compliance={activity.compliance.content()}
-              participantCount={activity.participantCount.content()}
-              latestActivity={activity.latestActivity.content()}
-              data-testid={`${dataTestid}-activity-card`}
-            />
-          ))}
-        </StyledFlexWrap>
-      )}
-
-      {isEmpty && <EmptyState>{getEmptyComponent()}</EmptyState>}
-    </StyledBody>
+    <ActivityGrid
+      rows={activities}
+      data-testid={dataTestid}
+      isLoading={isLoading}
+      order={'desc'}
+      orderBy={''}
+    />
   );
 };
